Separate installer from installation
