+++ internal/plugin/installer_test.go
